Refactor nautodog, add logging, and PrintInfo utility
This commit includes a major refactoring of the nautodog CLI tool,
the introduction of logging, and a new PrintInfo utility as per your request.

Key changes:
1.  **Project Structure Refactoring:**
    *   I moved the main script `naurodog.py` (renamed to `nautodog.py`) to the root directory.
    *   Command groups (`ddsnmpconfig`, `ddmonitor`, `ddmainconfig`, `ddagent`, `report`) were extracted into separate modules under `src/`. Each group now resides in `src/<group_name>/commands.py`.
    *   I created the `src/utils/` directory for utility modules.

2.  **Logging Implementation:**
    *   I integrated `loguru` for application logging, configured in the main `nautodog.py`.
    *   I created a `@log_command` decorator and applied it to all Click command functions. This decorator logs the command path and its parameters upon execution.

3.  **PrintInfo Utility:**
    *   I created `src/utils/print_info.py` containing a `PrintInfo` class.
    *   The `PrintInfo.print_caller_name()` method prints the name of the function that called it.
    *   This utility has been integrated into the `ddsnmpconfig` command group, where each command now calls `PrintInfo().print_caller_name()` at the beginning of its execution.

**Work In Progress:**
*   The integration of `PrintInfo().print_caller_name()` is complete for `ddsnmpconfig` but still needs to be applied to the `ddmonitor`, `ddmainconfig`, `ddagent`, and `report` command groups.
*   The `README.md` needs to be updated to reflect these changes and provide usage instructions.
*   Thorough testing of all commands, logging, and PrintInfo output is pending.

**Challenges:**
*   Initially, when I was working on file manipulation, especially writing multi-line Python code into new files, I encountered several syntax errors. I resolved this by executing a temporary Python script that performed the file I/O.

This represents a significant step towards the desired refactoring and feature additions. Next, I will focus on completing the `PrintInfo` integration, documentation, and testing.
