--- src/naurodog.py
@@ -1,164 +1,0 @@
-import click
-
-@click.group()
-def cli():
-    """Naurodog CLI application"""
-    pass
-
-# Group: ddsnmpconfig
-@cli.group()
-def ddsnmpconfig():
-    """Commands for SNMP configuration"""
-    pass
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def addsnmpv3(name):
-    click.echo(f"Command: addsnmpv3, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def addsnmpv2(name):
-    click.echo(f"Command: addsnmpv2, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def uploadcustopprofile(name): # Corrected typo: uploadcustopProfile -> uploadcustopprofile
-    click.echo(f"Command: uploadcustopprofile, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def createlocalconfig(name):
-    click.echo(f"Command: createlocalconfig, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def verifydevices(name):
-    click.echo(f"Command: verifydevices, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def autodiscoveryv2(name): # Corrected typo: autodicsoveryv2 -> autodiscoveryv2
-    click.echo(f"Command: autodiscoveryv2, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def autodiscoveryv3(name):
-    click.echo(f"Command: autodiscoveryv3, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def rollbackconfig(name):
-    click.echo(f"Command: rollbackconfig, Name: {name}")
-
-@ddsnmpconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def revokedevice(name): # Corrected typo: revokedivice -> revokedevice
-    click.echo(f"Command: revokedevice, Name: {name}")
-
-# Group: ddmonitor
-@cli.group()
-def ddmonitor():
-    """Commands for monitoring"""
-    pass
-
-@ddmonitor.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def addreachablemonitor(name):
-    click.echo(f"Command: addreachablemonitor, Name: {name}")
-
-@ddmonitor.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def addinterfacemonitor(name):
-    click.echo(f"Command: addinterfacemonitor, Name: {name}")
-
-@ddmonitor.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def addmonitorsbyrules(name):
-    click.echo(f"Command: addmonitorsbyrules, Name: {name}")
-
-# Group: ddmainconfig
-@cli.group()
-def ddmainconfig():
-    """Commands for main configuration"""
-    pass
-
-@ddmainconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def addtag(name):
-    click.echo(f"Command: addtag, Name: {name}")
-
-@ddmainconfig.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def apikey(name):
-    click.echo(f"Command: apikey, Name: {name}")
-
-# Group: ddagent
-@cli.group()
-def ddagent():
-    """Commands for ddagent"""
-    pass
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def ddastatus(name):
-    click.echo(f"Command: ddastatus, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def ddaerrors(name):
-    click.echo(f"Command: ddaerrors, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def ddalogs(name):
-    click.echo(f"Command: ddalogs, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def dduploadconfigs(name):
-    click.echo(f"Command: dduploadconfigs, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def dddownloadconfigs(name):
-    click.echo(f"Command: dddownloadconfigs, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def configconsistency(name):
-    click.echo(f"Command: configconsistency, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def downloadconfigs(name): # Note: This might be a duplicate or a more generic version of dddownloadconfigs
-    click.echo(f"Command: downloadconfigs, Name: {name}")
-
-@ddagent.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def uploadconfigs(name): # Note: This might be a duplicate or a more generic version of dduploadconfigs
-    click.echo(f"Command: uploadconfigs, Name: {name}")
-
-# Group: report
-@cli.group()
-def report():
-    """Commands for reporting"""
-    pass
-
-@report.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def ndreport_devices(name):
-    click.echo(f"Command: ndreport_devices, Name: {name}")
-
-@report.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def sldreport_devices(name):
-    click.echo(f"Command: sldreport_devices, Name: {name}")
-
-@report.command()
-@click.option('-n', '--name', required=True, help='Name to use.')
-def ddreportmonitorcoverage(name):
-    click.echo(f"Command: ddreportmonitorcoverage, Name: {name}")
-
-if __name__ == '__main__':
-    cli()
