+++ src/report/__init__.py
@@ -1,0 +1,1 @@
+# This is an empty __init__.py file
